+++ finalModule/finalModule/src/main/java/com/codegym/finalModule/repository/ISupplierRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ISupplierRepository extends JpaRepository<Supplier, Long> {
+public interface ISupplierRepository extends JpaRepository<Supplier, Integer> {
     Supplier findBySupplierCode(String supplierCode);
 }
