+++ finalModule/finalModule/src/main/java/com/codegym/finalModule/repository/IUserRoleRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface IUserRoleRepository extends JpaRepository<UserRole, Long> {
+public interface IUserRoleRepository extends JpaRepository<UserRole, Integer> {
     List<UserRole> findAllByUser(User user);
 }
